Demand-responsive transport planner

Tool for placing virtual stations in demand-responsive transport systems in villages by defining and minimizing a global energy (drtplanr, name is inspired by stplanr <https://github.com/ropensci/stplanr>). The station locations are randomly initialized in the street network and iteratively optimized based on the reachable population in combination with walking and driving times.

See also